EVG-15593 untangle service and graphql packages
