+++ graphql/integration_atomic_test.go
@@ -21,9 +21,10 @@ import (
 	"testing"
 	"time"
 
+	"github.com/evergreen-ci/evergreen/graphql"
+
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/db"
-	"github.com/evergreen-ci/evergreen/graphql"
 	"github.com/evergreen-ci/evergreen/model"
 	"github.com/evergreen-ci/evergreen/model/build"
 	"github.com/evergreen-ci/evergreen/model/commitqueue"
