+++ graphql/resolvers.go
@@ -4,12 +4,16 @@ import (
 	"context"
 	"fmt"
 	"net/http"
+	"net/url"
+	"runtime/debug"
 	"sort"
 	"strconv"
 	"strings"
 	"time"
 
 	"github.com/99designs/gqlgen/graphql"
+	"github.com/evergreen-ci/evergreen/model/version"
+
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/api"
 	"github.com/evergreen-ci/evergreen/apimodels"
@@ -412,7 +416,7 @@ func (r *taskResolver) DependsOn(ctx context.Context, at *restModel.APITask) ([]
 }
 
 func (r *taskResolver) CanOverrideDependencies(ctx context.Context, at *restModel.APITask) (bool, error) {
-	currentUser := MustHaveUser(ctx)
+	currentUser := mustHaveUser(ctx)
 	if at.OverrideDependencies {
 		return false, nil
 	}
@@ -437,7 +441,7 @@ func (r *projectResolver) IsFavorite(ctx context.Context, at *restModel.APIProje
 	if err != nil || p == nil {
 		return false, ResourceNotFound.Send(ctx, fmt.Sprintf("Could not find project: %s : %s", *at.Identifier, err))
 	}
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	if utility.StringSliceContains(usr.FavoriteProjects, *at.Identifier) {
 		return true, nil
 	}
@@ -567,7 +571,7 @@ func (r *mutationResolver) AddFavoriteProject(ctx context.Context, identifier st
 		return nil, ResourceNotFound.Send(ctx, fmt.Sprintf("could not find project '%s'", identifier))
 	}
 
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 
 	err = usr.AddFavoritedProject(identifier)
 	if err != nil {
@@ -587,7 +591,7 @@ func (r *mutationResolver) RemoveFavoriteProject(ctx context.Context, identifier
 		return nil, ResourceNotFound.Send(ctx, fmt.Sprintf("Could not find project: %s", identifier))
 	}
 
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 
 	err = usr.RemoveFavoriteProject(identifier)
 	if err != nil {
@@ -606,9 +610,15 @@ func (r *mutationResolver) SpawnVolume(ctx context.Context, spawnVolumeInput Spa
 	if err != nil {
 		return false, err
 	}
-	success, _, gqlErr, err, vol := RequestNewVolume(ctx, GetVolumeFromSpawnVolumeInput(spawnVolumeInput))
+	volumeRequest := host.Volume{
+		AvailabilityZone: spawnVolumeInput.AvailabilityZone,
+		Size:             spawnVolumeInput.Size,
+		Type:             spawnVolumeInput.Type,
+		CreatedBy:        mustHaveUser(ctx).Id,
+	}
+	vol, statusCode, err := cloud.RequestNewVolume(ctx, volumeRequest)
 	if err != nil {
-		return false, gqlErr.Send(ctx, err.Error())
+		return false, mapHTTPStatusToGqlError(ctx, statusCode, err)
 	}
 	if vol == nil {
 		return false, InternalServerError.Send(ctx, "Unable to create volume")
@@ -619,7 +629,7 @@ func (r *mutationResolver) SpawnVolume(ctx context.Context, spawnVolumeInput Spa
 		var newExpiration time.Time
 		newExpiration, err = restModel.FromTimePtr(spawnVolumeInput.Expiration)
 		if err != nil {
-			return false, gqlErr.Send(ctx, errors.Wrapf(err, errorTemplate, vol.ID).Error())
+			return false, InternalServerError.Send(ctx, errors.Wrapf(err, errorTemplate, vol.ID).Error())
 		}
 		additionalOptions.Expiration = newExpiration
 	} else if spawnVolumeInput.NoExpiration != nil && *spawnVolumeInput.NoExpiration == true {
@@ -631,13 +641,13 @@ func (r *mutationResolver) SpawnVolume(ctx context.Context, spawnVolumeInput Spa
 		return false, InternalServerError.Send(ctx, fmt.Sprintf("Unable to apply expiration options to volume %s: %s", vol.ID, err.Error()))
 	}
 	if spawnVolumeInput.Host != nil {
-		_, _, gqlErr, err := AttachVolume(ctx, vol.ID, *spawnVolumeInput.Host)
+		statusCode, err := cloud.AttachVolume(ctx, vol.ID, *spawnVolumeInput.Host)
 		if err != nil {
-			return false, gqlErr.Send(ctx, errors.Wrapf(err, errorTemplate, vol.ID).Error())
+			return false, mapHTTPStatusToGqlError(ctx, statusCode, errors.Wrapf(err, errorTemplate, vol.ID))
 		}
 	}
 
-	return success, nil
+	return true, nil
 }
 
 func (r *mutationResolver) UpdateVolume(ctx context.Context, updateVolumeInput UpdateVolumeInput) (bool, error) {
@@ -686,7 +696,7 @@ func (r *mutationResolver) UpdateVolume(ctx context.Context, updateVolumeInput U
 }
 
 func (r *mutationResolver) SpawnHost(ctx context.Context, spawnHostInput *SpawnHostInput) (*restModel.APIHost, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	if spawnHostInput.SavePublicKey {
 		if err := savePublicKey(ctx, *spawnHostInput.PublicKey); err != nil {
 			return nil, err
@@ -772,13 +782,13 @@ func (r *mutationResolver) SpawnHost(ctx context.Context, spawnHostInput *SpawnH
 
 func (r *mutationResolver) EditSpawnHost(ctx context.Context, editSpawnHostInput *EditSpawnHostInput) (*restModel.APIHost, error) {
 	var v *host.Volume
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	h, err := host.FindOneByIdOrTag(editSpawnHostInput.HostID)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error finding host by id: %s", err))
 	}
 
-	if !CanUpdateSpawnHost(h, usr) {
+	if !host.CanUpdateSpawnHost(h, usr) {
 		return nil, Forbidden.Send(ctx, "You are not authorized to modify this host")
 	}
 
@@ -850,7 +860,7 @@ func (r *mutationResolver) EditSpawnHost(ctx context.Context, editSpawnHostInput
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error modifying spawn host: %s", err))
 	}
 	if editSpawnHostInput.ServicePassword != nil {
-		_, _, err = UpdateHostPassword(ctx, evergreen.GetEnvironment(), h, usr, *editSpawnHostInput.ServicePassword, nil)
+		_, err = cloud.SetHostRDPPassword(ctx, evergreen.GetEnvironment(), h, *editSpawnHostInput.ServicePassword)
 		if err != nil {
 			return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error setting spawn host password: %s", err))
 		}
@@ -865,54 +875,48 @@ func (r *mutationResolver) EditSpawnHost(ctx context.Context, editSpawnHostInput
 }
 
 func (r *mutationResolver) UpdateSpawnHostStatus(ctx context.Context, hostID string, action SpawnHostStatusActions) (*restModel.APIHost, error) {
-	host, err := host.FindOneByIdOrTag(hostID)
+	h, err := host.FindOneByIdOrTag(hostID)
 	if err != nil {
 		return nil, ResourceNotFound.Send(ctx, fmt.Sprintf("Error finding host by id: %s", err))
 	}
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	env := evergreen.GetEnvironment()
 
-	if !CanUpdateSpawnHost(host, usr) {
+	if !host.CanUpdateSpawnHost(h, usr) {
 		return nil, Forbidden.Send(ctx, "You are not authorized to modify this host")
 	}
 
+	var httpStatus int
 	switch action {
 	case SpawnHostStatusActionsStart:
-		h, httpStatus, err := StartSpawnHost(ctx, env, host, usr, nil)
-		if err != nil {
-			return nil, mapHTTPStatusToGqlError(ctx, httpStatus, err)
-		}
-		apiHost := restModel.APIHost{}
-		err = apiHost.BuildFromService(h)
-		if err != nil {
-			return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error building apiHost from service: %s", err))
-		}
-		return &apiHost, nil
+		httpStatus, err = data.StartSpawnHost(ctx, env, usr, h)
 	case SpawnHostStatusActionsStop:
-		h, httpStatus, err := StopSpawnHost(ctx, env, host, usr, nil)
-		if err != nil {
-			return nil, mapHTTPStatusToGqlError(ctx, httpStatus, err)
-		}
-		apiHost := restModel.APIHost{}
-		err = apiHost.BuildFromService(h)
-		if err != nil {
-			return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error building apiHost from service: %s", err))
-		}
-		return &apiHost, nil
+		httpStatus, err = data.StopSpawnHost(ctx, env, usr, h)
 	case SpawnHostStatusActionsTerminate:
-		h, httpStatus, err := TerminateSpawnHost(ctx, env, host, usr, nil)
-		if err != nil {
-			return nil, mapHTTPStatusToGqlError(ctx, httpStatus, err)
-		}
-		apiHost := restModel.APIHost{}
-		err = apiHost.BuildFromService(h)
-		if err != nil {
-			return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error building apiHost from service: %s", err))
-		}
-		return &apiHost, nil
+		httpStatus, err = data.TerminateSpawnHost(ctx, env, usr, h)
 	default:
 		return nil, ResourceNotFound.Send(ctx, fmt.Sprintf("Could not find matching status for action : %s", action))
 	}
+	if err != nil {
+		if httpStatus == http.StatusInternalServerError {
+			var parsedUrl, _ = url.Parse("/graphql/query")
+			grip.Error(message.WrapError(err, message.Fields{
+				"method":  "POST",
+				"url":     parsedUrl,
+				"code":    httpStatus,
+				"action":  action,
+				"request": gimlet.GetRequestID(ctx),
+				"stack":   string(debug.Stack()),
+			}))
+		}
+		return nil, mapHTTPStatusToGqlError(ctx, httpStatus, err)
+	}
+	apiHost := restModel.APIHost{}
+	err = apiHost.BuildFromService(h)
+	if err != nil {
+		return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error building apiHost from service: %s", err))
+	}
+	return &apiHost, nil
 
 }
 
@@ -1039,19 +1043,16 @@ func (r *queryResolver) Host(ctx context.Context, hostID string) (*restModel.API
 }
 
 func (r *queryResolver) MyVolumes(ctx context.Context) ([]*restModel.APIVolume, error) {
-	volumes, err := GetMyVolumes(MustHaveUser(ctx))
+	usr := mustHaveUser(ctx)
+	volumes, err := host.FindSortedVolumesByUser(usr.Username())
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, err.Error())
 	}
-	volumePointers := make([]*restModel.APIVolume, 0, len(volumes))
-	for i, _ := range volumes {
-		volumePointers = append(volumePointers, &volumes[i])
-	}
-	return volumePointers, nil
+	return getAPIVolumeList(volumes)
 }
 
 func (r *queryResolver) MyHosts(ctx context.Context) ([]*restModel.APIHost, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	hosts, err := host.Find(host.ByUserWithRunningStatus(usr.Username()))
 	if err != nil {
 		return nil, InternalServerError.Send(ctx,
@@ -1192,7 +1193,7 @@ func (r *mutationResolver) CopyProject(ctx context.Context, opts data.CopyProjec
 }
 
 func (r *mutationResolver) AttachProjectToNewRepo(ctx context.Context, obj MoveProjectInput) (*restModel.APIProjectRef, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 
 	pRef, err := r.sc.FindProjectById(obj.ProjectID, false, false)
 	if err != nil || pRef == nil {
@@ -1213,19 +1214,19 @@ func (r *mutationResolver) AttachProjectToNewRepo(ctx context.Context, obj MoveP
 }
 
 func (r *mutationResolver) AttachVolumeToHost(ctx context.Context, volumeAndHost VolumeHost) (bool, error) {
-	success, _, gqlErr, err := AttachVolume(ctx, volumeAndHost.VolumeID, volumeAndHost.HostID)
+	statusCode, err := cloud.AttachVolume(ctx, volumeAndHost.VolumeID, volumeAndHost.HostID)
 	if err != nil {
-		return false, gqlErr.Send(ctx, err.Error())
+		return false, mapHTTPStatusToGqlError(ctx, statusCode, err)
 	}
-	return success, nil
+	return statusCode == http.StatusOK, nil
 }
 
 func (r *mutationResolver) DetachVolumeFromHost(ctx context.Context, volumeID string) (bool, error) {
-	success, _, gqlErr, err := DetachVolume(ctx, volumeID)
+	statusCode, err := cloud.DetachVolume(ctx, volumeID)
 	if err != nil {
-		return false, gqlErr.Send(ctx, err.Error())
+		return false, mapHTTPStatusToGqlError(ctx, statusCode, err)
 	}
-	return success, nil
+	return statusCode == http.StatusOK, nil
 }
 
 type patchResolver struct{ *Resolver }
@@ -1352,17 +1353,17 @@ func (r *patchResolver) Duration(ctx context.Context, obj *restModel.APIPatch) (
 }
 
 func (r *patchResolver) Time(ctx context.Context, obj *restModel.APIPatch) (*PatchTime, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 
-	started, err := GetFormattedDate(obj.StartTime, usr.Settings.Timezone)
+	started, err := getFormattedDate(obj.StartTime, usr.Settings.Timezone)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, err.Error())
 	}
-	finished, err := GetFormattedDate(obj.FinishTime, usr.Settings.Timezone)
+	finished, err := getFormattedDate(obj.FinishTime, usr.Settings.Timezone)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, err.Error())
 	}
-	submittedAt, err := GetFormattedDate(obj.CreateTime, usr.Settings.Timezone)
+	submittedAt, err := getFormattedDate(obj.CreateTime, usr.Settings.Timezone)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, err.Error())
 	}
@@ -1391,7 +1392,7 @@ func (r *patchResolver) BaseVersionID(ctx context.Context, obj *restModel.APIPat
 }
 
 func (r *patchResolver) Project(ctx context.Context, apiPatch *restModel.APIPatch) (*PatchProject, error) {
-	patchProject, err := GetPatchProjectVariantsAndTasksForUI(ctx, apiPatch)
+	patchProject, err := getPatchProjectVariantsAndTasksForUI(ctx, apiPatch)
 	if err != nil {
 		return nil, err
 	}
@@ -1548,7 +1549,7 @@ func (r *projectResolver) Patches(ctx context.Context, obj *restModel.APIProject
 }
 
 func (r *queryResolver) UserSettings(ctx context.Context) (*restModel.APIUserSettings, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	userSettings := restModel.APIUserSettings{}
 	err := userSettings.BuildFromService(usr.Settings)
 	if err != nil {
@@ -1565,7 +1566,7 @@ func (r *queryResolver) Task(ctx context.Context, taskID string, execution *int)
 	if dbTask == nil {
 		return nil, errors.Errorf("unable to find task %s", taskID)
 	}
-	apiTask, err := GetAPITaskFromTask(ctx, r.sc, *dbTask)
+	apiTask, err := getAPITaskFromTask(ctx, r.sc, *dbTask)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, "error converting task")
 	}
@@ -1592,13 +1593,13 @@ func (r *queryResolver) TaskAllExecutions(ctx context.Context, taskID string) ([
 			return nil, errors.Errorf("unable to find task %s", taskID)
 		}
 		var apiTask *restModel.APITask
-		apiTask, err = GetAPITaskFromTask(ctx, r.sc, *dbTask)
+		apiTask, err = getAPITaskFromTask(ctx, r.sc, *dbTask)
 		if err != nil {
 			return nil, InternalServerError.Send(ctx, "error converting task")
 		}
 		allTasks = append(allTasks, apiTask)
 	}
-	apiTask, err := GetAPITaskFromTask(ctx, r.sc, *latestTask)
+	apiTask, err := getAPITaskFromTask(ctx, r.sc, *latestTask)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, "error converting task")
 	}
@@ -1618,7 +1619,7 @@ func (r *queryResolver) Projects(ctx context.Context) ([]*GroupedProjects, error
 			enabledProjects = append(enabledProjects, p)
 		}
 	}
-	groupedProjects, err := GroupProjects(enabledProjects, false)
+	groupedProjects, err := groupProjects(enabledProjects, false)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error grouping project: %s", err.Error()))
 	}
@@ -1626,7 +1627,7 @@ func (r *queryResolver) Projects(ctx context.Context) ([]*GroupedProjects, error
 }
 
 func (r *queryResolver) ViewableProjectRefs(ctx context.Context) ([]*GroupedProjects, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	projectIds, err := usr.GetViewableProjectSettings()
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error getting viewable projects for '%s': '%s'", usr.DispName, err.Error()))
@@ -1637,7 +1638,7 @@ func (r *queryResolver) ViewableProjectRefs(ctx context.Context) ([]*GroupedProj
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error getting projects: %v", err.Error()))
 	}
 
-	groupedProjects, err := GroupProjects(projects, true)
+	groupedProjects, err := groupProjects(projects, true)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error grouping project: %s", err.Error()))
 	}
@@ -1963,7 +1964,7 @@ func (r *queryResolver) TaskFiles(ctx context.Context, taskID string, execution 
 			return &emptyTaskFiles, ResourceNotFound.Send(ctx, err.Error())
 		}
 		for _, execTask := range execTasks {
-			groupedFiles, err := GetGroupedFiles(ctx, execTask.DisplayName, execTask.Id, t.Execution)
+			groupedFiles, err := getGroupedFiles(ctx, execTask.DisplayName, execTask.Id, t.Execution)
 			if err != nil {
 				return &emptyTaskFiles, err
 			}
@@ -1971,7 +1972,7 @@ func (r *queryResolver) TaskFiles(ctx context.Context, taskID string, execution 
 			groupedFilesList = append(groupedFilesList, groupedFiles)
 		}
 	} else {
-		groupedFiles, err := GetGroupedFiles(ctx, t.DisplayName, taskID, t.Execution)
+		groupedFiles, err := getGroupedFiles(ctx, t.DisplayName, taskID, t.Execution)
 		if err != nil {
 			return &emptyTaskFiles, err
 		}
@@ -2262,7 +2263,7 @@ func (r *queryResolver) CommitQueue(ctx context.Context, id string) (*restModel.
 }
 
 func (r *queryResolver) UserConfig(ctx context.Context) (*UserConfig, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	settings := evergreen.GetEnvironment().Settings()
 	config := &UserConfig{
 		User:          usr.Username(),
@@ -2429,7 +2430,7 @@ func (r *taskQueueItemResolver) Requester(ctx context.Context, obj *restModel.AP
 
 func (r *mutationResolver) SaveProjectSettingsForSection(ctx context.Context, obj *restModel.APIProjectSettings, section ProjectSettingsSection) (*restModel.APIProjectSettings, error) {
 	projectId := utility.FromStringPtr(obj.ProjectRef.Id)
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	changes, err := r.sc.SaveProjectSettingsForSection(ctx, projectId, obj, model.ProjectPageSection(section), false, usr.Username())
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error saving project settings for section: %s", err.Error()))
@@ -2439,7 +2440,7 @@ func (r *mutationResolver) SaveProjectSettingsForSection(ctx context.Context, ob
 
 func (r *mutationResolver) SaveRepoSettingsForSection(ctx context.Context, obj *restModel.APIProjectSettings, section ProjectSettingsSection) (*restModel.APIProjectSettings, error) {
 	projectId := utility.FromStringPtr(obj.ProjectRef.Id)
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	changes, err := r.sc.SaveProjectSettingsForSection(ctx, projectId, obj, model.ProjectPageSection(section), true, usr.Username())
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error saving project settings for section: %s", err.Error()))
@@ -2448,7 +2449,7 @@ func (r *mutationResolver) SaveRepoSettingsForSection(ctx context.Context, obj *
 }
 
 func (r *mutationResolver) DefaultSectionToRepo(ctx context.Context, projectId string, section ProjectSettingsSection) (*string, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	if err := model.DefaultSectionToRepo(projectId, model.ProjectPageSection(section), usr.Username()); err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error defaulting to repo for section: %s", err.Error()))
 	}
@@ -2457,7 +2458,7 @@ func (r *mutationResolver) DefaultSectionToRepo(ctx context.Context, projectId s
 }
 
 func (r *mutationResolver) AttachProjectToRepo(ctx context.Context, projectID string) (*restModel.APIProjectRef, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	pRef, err := r.sc.FindProjectById(projectID, false, false)
 	if err != nil {
 		return nil, ResourceNotFound.Send(ctx, fmt.Sprintf("error finding project %s: %s", projectID, err.Error()))
@@ -2477,7 +2478,7 @@ func (r *mutationResolver) AttachProjectToRepo(ctx context.Context, projectID st
 }
 
 func (r *mutationResolver) DetachProjectFromRepo(ctx context.Context, projectID string) (*restModel.APIProjectRef, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	pRef, err := r.sc.FindProjectById(projectID, false, false)
 	if err != nil {
 		return nil, ResourceNotFound.Send(ctx, fmt.Sprintf("error finding project %s: %s", projectID, err.Error()))
@@ -2537,12 +2538,12 @@ func (r *mutationResolver) SetTaskPriority(ctx context.Context, taskID string, p
 	if t == nil {
 		return nil, ResourceNotFound.Send(ctx, fmt.Sprintf("cannot find task with id %s", taskID))
 	}
-	apiTask, err := GetAPITaskFromTask(ctx, r.sc, *t)
+	apiTask, err := getAPITaskFromTask(ctx, r.sc, *t)
 	return apiTask, err
 }
 
 func (r *mutationResolver) OverrideTaskDependencies(ctx context.Context, taskID string) (*restModel.APITask, error) {
-	currentUser := MustHaveUser(ctx)
+	currentUser := mustHaveUser(ctx)
 	t, err := task.FindByIdExecution(taskID, nil)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error finding task %s: %s", taskID, err.Error()))
@@ -2554,12 +2555,11 @@ func (r *mutationResolver) OverrideTaskDependencies(ctx context.Context, taskID 
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error overriding dependencies for task %s: %s", taskID, err.Error()))
 	}
 	t.DisplayStatus = t.GetDisplayStatus()
-	return GetAPITaskFromTask(ctx, r.sc, *t)
+	return getAPITaskFromTask(ctx, r.sc, *t)
 }
 
 func (r *mutationResolver) SchedulePatch(ctx context.Context, patchID string, configure PatchConfigure) (*restModel.APIPatch, error) {
-	patchUpdateReq := PatchUpdate{}
-	patchUpdateReq.BuildFromGqlInput(configure)
+	patchUpdateReq := buildFromGqlInput(configure)
 	version, err := r.sc.FindVersionById(patchID)
 	if err != nil {
 		// FindVersionById does not distinguish between nil version err and db err; therefore must check that err
@@ -2568,11 +2568,11 @@ func (r *mutationResolver) SchedulePatch(ctx context.Context, patchID string, co
 			return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error occurred fetching patch `%s`: %s", patchID, err.Error()))
 		}
 	}
-	err, _, _, versionID := SchedulePatch(ctx, patchID, version, patchUpdateReq)
+	statusCode, err := units.SchedulePatch(ctx, patchID, version, patchUpdateReq)
 	if err != nil {
-		return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error scheduling patch `%s`: %s", patchID, err))
+		return nil, mapHTTPStatusToGqlError(ctx, statusCode, errors.Errorf("Error scheduling patch `%s`: %s", patchID, err.Error()))
 	}
-	scheduledPatch, err := r.sc.FindPatchById(versionID)
+	scheduledPatch, err := r.sc.FindPatchById(patchID)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error getting scheduled patch `%s`: %s", patchID, err))
 	}
@@ -2580,12 +2580,12 @@ func (r *mutationResolver) SchedulePatch(ctx context.Context, patchID string, co
 }
 
 func (r *mutationResolver) SchedulePatchTasks(ctx context.Context, patchID string) (*string, error) {
-	modifications := VersionModifications{
-		Action: SetActive,
+	modifications := version.Modification{
+		Action: version.SetActive,
 		Active: true,
 		Abort:  false,
 	}
-	err := ModifyVersionHandler(ctx, r.sc, patchID, modifications)
+	err := modifyVersionHandler(ctx, r.sc, patchID, modifications)
 	if err != nil {
 		return nil, err
 	}
@@ -2593,12 +2593,12 @@ func (r *mutationResolver) SchedulePatchTasks(ctx context.Context, patchID strin
 }
 
 func (r *mutationResolver) UnschedulePatchTasks(ctx context.Context, patchID string, abort bool) (*string, error) {
-	modifications := VersionModifications{
-		Action: SetActive,
+	modifications := version.Modification{
+		Action: version.SetActive,
 		Active: false,
 		Abort:  abort,
 	}
-	err := ModifyVersionHandler(ctx, r.sc, patchID, modifications)
+	err := modifyVersionHandler(ctx, r.sc, patchID, modifications)
 	if err != nil {
 		return nil, err
 	}
@@ -2658,7 +2658,7 @@ func (r *mutationResolver) ScheduleUndispatchedBaseTasks(ctx context.Context, pa
 	}
 
 	for taskId := range tasksToSchedule {
-		task, err := SetScheduled(ctx, r.sc, taskId, true)
+		task, err := setScheduled(ctx, r.sc, taskId, true)
 		if err != nil {
 			return nil, err
 		}
@@ -2676,12 +2676,12 @@ func (r *mutationResolver) RestartPatch(ctx context.Context, patchID string, abo
 	if len(taskIds) == 0 {
 		return nil, InputValidationError.Send(ctx, "`taskIds` array is empty. You must provide at least one task id")
 	}
-	modifications := VersionModifications{
-		Action:  Restart,
+	modifications := version.Modification{
+		Action:  version.Restart,
 		Abort:   abort,
 		TaskIds: taskIds,
 	}
-	err := ModifyVersionHandler(ctx, r.sc, patchID, modifications)
+	err := modifyVersionHandler(ctx, r.sc, patchID, modifications)
 	if err != nil {
 		return nil, err
 	}
@@ -2692,12 +2692,12 @@ func (r *mutationResolver) RestartVersions(ctx context.Context, patchID string, 
 	if len(versionsToRestart) == 0 {
 		return nil, InputValidationError.Send(ctx, "No versions provided. You must provide at least one version to restart")
 	}
-	modifications := VersionModifications{
-		Action:            Restart,
+	modifications := version.Modification{
+		Action:            version.Restart,
 		Abort:             abort,
 		VersionsToRestart: versionsToRestart,
 	}
-	err := ModifyVersionHandler(ctx, r.sc, patchID, modifications)
+	err := modifyVersionHandler(ctx, r.sc, patchID, modifications)
 	if err != nil {
 		return nil, err
 	}
@@ -2723,11 +2723,11 @@ func (r *mutationResolver) RestartVersions(ctx context.Context, patchID string, 
 }
 
 func (r *mutationResolver) SetPatchPriority(ctx context.Context, patchID string, priority int) (*string, error) {
-	modifications := VersionModifications{
-		Action:   SetPriority,
+	modifications := version.Modification{
+		Action:   version.SetPriority,
 		Priority: int64(priority),
 	}
-	err := ModifyVersionHandler(ctx, r.sc, patchID, modifications)
+	err := modifyVersionHandler(ctx, r.sc, patchID, modifications)
 	if err != nil {
 		return nil, err
 	}
@@ -2735,7 +2735,7 @@ func (r *mutationResolver) SetPatchPriority(ctx context.Context, patchID string,
 }
 
 func (r *mutationResolver) EnqueuePatch(ctx context.Context, patchID string, commitMessage *string) (*restModel.APIPatch, error) {
-	user := MustHaveUser(ctx)
+	user := mustHaveUser(ctx)
 
 	existingPatch, err := r.sc.FindPatchById(patchID)
 	if err != nil {
@@ -2774,7 +2774,7 @@ func (r *mutationResolver) ScheduleTasks(ctx context.Context, taskIds []string) 
 	scheduledTasks := []*restModel.APITask{}
 	count := 0
 	for _, taskId := range taskIds {
-		task, err := SetScheduled(ctx, r.sc, taskId, true)
+		task, err := setScheduled(ctx, r.sc, taskId, true)
 		if err != nil {
 			return scheduledTasks, InternalServerError.Send(ctx, fmt.Sprintf("Failed to schedule %d task : %s", len(taskIds)-count, err.Error()))
 		}
@@ -2784,7 +2784,7 @@ func (r *mutationResolver) ScheduleTasks(ctx context.Context, taskIds []string) 
 	return scheduledTasks, nil
 }
 func (r *mutationResolver) ScheduleTask(ctx context.Context, taskID string) (*restModel.APITask, error) {
-	task, err := SetScheduled(ctx, r.sc, taskID, true)
+	task, err := setScheduled(ctx, r.sc, taskID, true)
 	if err != nil {
 		return nil, err
 	}
@@ -2792,7 +2792,7 @@ func (r *mutationResolver) ScheduleTask(ctx context.Context, taskID string) (*re
 }
 
 func (r *mutationResolver) UnscheduleTask(ctx context.Context, taskID string) (*restModel.APITask, error) {
-	task, err := SetScheduled(ctx, r.sc, taskID, false)
+	task, err := setScheduled(ctx, r.sc, taskID, false)
 	if err != nil {
 		return nil, err
 	}
@@ -2819,12 +2819,12 @@ func (r *mutationResolver) AbortTask(ctx context.Context, taskID string) (*restM
 	if t == nil {
 		return nil, ResourceNotFound.Send(ctx, fmt.Sprintf("cannot find task with id %s", taskID))
 	}
-	apiTask, err := GetAPITaskFromTask(ctx, r.sc, *t)
+	apiTask, err := getAPITaskFromTask(ctx, r.sc, *t)
 	return apiTask, err
 }
 
 func (r *mutationResolver) RestartTask(ctx context.Context, taskID string) (*restModel.APITask, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	username := usr.Username()
 	if err := model.TryResetTask(taskID, username, evergreen.UIPackage, nil); err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("error restarting task %s: %s", taskID, err.Error()))
@@ -2836,13 +2836,13 @@ func (r *mutationResolver) RestartTask(ctx context.Context, taskID string) (*res
 	if t == nil {
 		return nil, ResourceNotFound.Send(ctx, fmt.Sprintf("cannot find task with id %s", taskID))
 	}
-	apiTask, err := GetAPITaskFromTask(ctx, r.sc, *t)
+	apiTask, err := getAPITaskFromTask(ctx, r.sc, *t)
 	return apiTask, err
 }
 
 // EditAnnotationNote updates the note for the annotation, assuming it hasn't been updated in the meantime.
 func (r *mutationResolver) EditAnnotationNote(ctx context.Context, taskID string, execution int, originalMessage, newMessage string) (bool, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	if err := annotations.UpdateAnnotationNote(taskID, execution, originalMessage, newMessage, usr.Username()); err != nil {
 		return false, InternalServerError.Send(ctx, fmt.Sprintf("couldn't update note: %s", err.Error()))
 	}
@@ -2852,7 +2852,7 @@ func (r *mutationResolver) EditAnnotationNote(ctx context.Context, taskID string
 // MoveAnnotationIssue moves an issue for the annotation. If isIssue is set, it removes the issue from Issues and adds it
 // to Suspected Issues, otherwise vice versa.
 func (r *mutationResolver) MoveAnnotationIssue(ctx context.Context, taskID string, execution int, apiIssue restModel.APIIssueLink, isIssue bool) (bool, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	issue := restModel.APIIssueLinkToService(apiIssue)
 	if isIssue {
 		if err := annotations.MoveIssueToSuspectedIssue(taskID, execution, *issue, usr.Username()); err != nil {
@@ -2871,7 +2871,7 @@ func (r *mutationResolver) MoveAnnotationIssue(ctx context.Context, taskID strin
 // If isIssue is set, it adds to Issues, otherwise it adds to Suspected Issues.
 func (r *mutationResolver) AddAnnotationIssue(ctx context.Context, taskID string, execution int,
 	apiIssue restModel.APIIssueLink, isIssue bool) (bool, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	issue := restModel.APIIssueLinkToService(apiIssue)
 	if err := util.CheckURL(issue.URL); err != nil {
 		return false, InputValidationError.Send(ctx, fmt.Sprintf("issue does not have valid URL: %s", err.Error()))
@@ -2921,7 +2921,7 @@ func (r *mutationResolver) RemoveItemFromCommitQueue(ctx context.Context, commit
 }
 
 func (r *mutationResolver) ClearMySubscriptions(ctx context.Context) (int, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	username := usr.Username()
 	subs, err := r.sc.GetSubscriptions(username, event.OwnerTypePerson)
 	if err != nil {
@@ -2942,7 +2942,7 @@ func (r *mutationResolver) ClearMySubscriptions(ctx context.Context) (int, error
 }
 
 func (r *mutationResolver) SaveSubscription(ctx context.Context, subscription restModel.APISubscription) (bool, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	username := usr.Username()
 	idType, id, err := getResourceTypeAndIdFromSubscriptionSelectors(ctx, subscription.Selectors)
 	if err != nil {
@@ -2992,7 +2992,7 @@ func (r *mutationResolver) SaveSubscription(ctx context.Context, subscription re
 }
 
 func (r *mutationResolver) UpdateUserSettings(ctx context.Context, userSettings *restModel.APIUserSettings) (bool, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 
 	updatedUserSettings, err := restModel.UpdateUserSettings(ctx, usr, *userSettings)
 	if err != nil {
@@ -3006,7 +3006,7 @@ func (r *mutationResolver) UpdateUserSettings(ctx context.Context, userSettings 
 }
 
 func (r *mutationResolver) RestartJasper(ctx context.Context, hostIds []string) (int, error) {
-	user := MustHaveUser(ctx)
+	user := mustHaveUser(ctx)
 
 	hosts, permissions, httpStatus, err := api.GetHostsAndUserPermissions(user, hostIds)
 	if err != nil {
@@ -3022,7 +3022,7 @@ func (r *mutationResolver) RestartJasper(ctx context.Context, hostIds []string) 
 }
 
 func (r *mutationResolver) ReprovisionToNew(ctx context.Context, hostIds []string) (int, error) {
-	user := MustHaveUser(ctx)
+	user := mustHaveUser(ctx)
 
 	hosts, permissions, httpStatus, err := api.GetHostsAndUserPermissions(user, hostIds)
 	if err != nil {
@@ -3038,7 +3038,7 @@ func (r *mutationResolver) ReprovisionToNew(ctx context.Context, hostIds []strin
 }
 
 func (r *mutationResolver) UpdateHostStatus(ctx context.Context, hostIds []string, status string, notes *string) (int, error) {
-	user := MustHaveUser(ctx)
+	user := mustHaveUser(ctx)
 
 	hosts, permissions, httpStatus, err := api.GetHostsAndUserPermissions(user, hostIds)
 	if err != nil {
@@ -3068,7 +3068,7 @@ func (r *mutationResolver) RemovePublicKey(ctx context.Context, keyName string) 
 	if !doesPublicKeyNameAlreadyExist(ctx, keyName) {
 		return nil, InputValidationError.Send(ctx, fmt.Sprintf("Error deleting public key. Provided key name, %s, does not exist.", keyName))
 	}
-	err := MustHaveUser(ctx).DeletePublicKey(keyName)
+	err := mustHaveUser(ctx).DeletePublicKey(keyName)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error deleting public key: %s", err.Error()))
 	}
@@ -3077,11 +3077,11 @@ func (r *mutationResolver) RemovePublicKey(ctx context.Context, keyName string) 
 }
 
 func (r *mutationResolver) RemoveVolume(ctx context.Context, volumeID string) (bool, error) {
-	success, _, gqlErr, err := DeleteVolume(ctx, volumeID)
+	statusCode, err := cloud.DeleteVolume(ctx, volumeID)
 	if err != nil {
-		return false, gqlErr.Send(ctx, err.Error())
+		return false, mapHTTPStatusToGqlError(ctx, statusCode, err)
 	}
-	return success, nil
+	return statusCode == http.StatusOK, nil
 }
 
 func (r *mutationResolver) UpdatePublicKey(ctx context.Context, targetKeyName string, updateInfo PublicKeyInput) ([]*restModel.APIPubKey, error) {
@@ -3095,7 +3095,7 @@ func (r *mutationResolver) UpdatePublicKey(ctx context.Context, targetKeyName st
 	if err != nil {
 		return nil, err
 	}
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	err = usr.UpdatePublicKey(targetKeyName, updateInfo.Name, updateInfo.Key)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, fmt.Sprintf("Error updating public key, %s: %s", targetKeyName, err.Error()))
@@ -3105,7 +3105,7 @@ func (r *mutationResolver) UpdatePublicKey(ctx context.Context, targetKeyName st
 }
 
 func (r *queryResolver) User(ctx context.Context, userIdParam *string) (*restModel.APIDBUser, error) {
-	usr := MustHaveUser(ctx)
+	usr := mustHaveUser(ctx)
 	var err error
 	if userIdParam != nil {
 		usr, err = user.FindOneById(*userIdParam)
@@ -3508,7 +3508,7 @@ func (r *taskResolver) Patch(ctx context.Context, obj *restModel.APITask) (*rest
 func (r *queryResolver) BuildBaron(ctx context.Context, taskID string, exec int) (*BuildBaron, error) {
 	execString := strconv.Itoa(exec)
 
-	searchReturnInfo, bbConfig, err := GetSearchReturnInfo(taskID, execString)
+	searchReturnInfo, bbConfig, err := model.GetSearchReturnInfo(taskID, execString)
 	if err != nil {
 		return nil, InternalServerError.Send(ctx, err.Error())
 	}
@@ -3516,26 +3516,21 @@ func (r *queryResolver) BuildBaron(ctx context.Context, taskID string, exec int)
 	return &BuildBaron{
 		SearchReturnInfo:        searchReturnInfo,
 		BuildBaronConfigured:    bbConfig.ProjectFound && bbConfig.SearchConfigured,
-		BbTicketCreationDefined: bbConfig.ticketCreationDefined,
+		BbTicketCreationDefined: bbConfig.TicketCreationDefined,
 	}, nil
 }
 
 func (r *mutationResolver) BbCreateTicket(ctx context.Context, taskID string, execution *int) (bool, error) {
-	taskNotFound, err := BbFileTicket(ctx, taskID, *execution)
-	successful := true
-
+	httpStatus, err := data.BbFileTicket(ctx, taskID, *execution)
 	if err != nil {
-		return !successful, InternalServerError.Send(ctx, err.Error())
-	}
-	if taskNotFound {
-		return !successful, ResourceNotFound.Send(ctx, fmt.Sprintf("could not find task '%s'", taskID))
+		return false, mapHTTPStatusToGqlError(ctx, httpStatus, err)
 	}
 
-	return successful, nil
+	return true, nil
 }
 
 func (r *queryResolver) BbGetCreatedTickets(ctx context.Context, taskID string) ([]*thirdparty.JiraTicket, error) {
-	createdTickets, err := BbGetCreatedTicketsPointers(taskID)
+	createdTickets, err := bbGetCreatedTicketsPointers(taskID)
 	if err != nil {
 		return nil, err
 	}
