+++ service/version.go
@@ -5,9 +5,10 @@ import (
 	"net/http"
 	"sort"
 
+	"github.com/evergreen-ci/evergreen/model/version"
+
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/db"
-	"github.com/evergreen-ci/evergreen/graphql"
 	"github.com/evergreen-ci/evergreen/model"
 	"github.com/evergreen-ci/evergreen/model/build"
 	"github.com/evergreen-ci/evergreen/model/task"
@@ -251,12 +252,12 @@ func (uis *UIServer) modifyVersion(w http.ResponseWriter, r *http.Request) {
 	}
 	user := MustHaveUser(r)
 
-	modifications := graphql.VersionModifications{}
-	if err = utility.ReadJSON(utility.NewRequestReader(r), &modifications); err != nil {
+	modification := version.Modification{}
+	if err = utility.ReadJSON(utility.NewRequestReader(r), &modification); err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
-	httpStatus, err := graphql.ModifyVersion(*projCtx.Version, *user, modifications)
+	httpStatus, err := version.ModifyVersion(*projCtx.Version, *user, modification)
 	if err != nil {
 		http.Error(w, err.Error(), httpStatus)
 		return
